refactor(store-core): extract environment handlers into handlers/environment.ts (audit P2-3)
Third slice of the P2-3 handlers/index.ts split. Move the two environment
message handlers (environment_list / environment_error) into a new
handlers/environment.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). Fully self-contained — no shared barrel
helpers and no ../types imports (EnvironmentInfo lives downstream). The
barrel re-exports via `export * from './environment'`, so the public
surface is unchanged and every consumer is unaffected.

handlers/index.ts: 5252 -> 5209 lines.
